flux models folder #71
