+++ src/flux/BurnupHeatFluxModel.cpp
@@ -6,8 +6,8 @@
  * @author Jean‑Baptiste Filippi — 2025
  */
 
-#include "FluxModel.h"
-#include "FireDomain.h"
+#include "../FluxModel.h"
+#include "../FireDomain.h"
 #include <math.h>
 
 using namespace std;
