+++ src/flux/LavaHeatFluxModel.cpp
@@ -6,9 +6,9 @@
  * @author Jean‑Baptiste Filippi — 2025
  */
 
-#include "FluxModel.h"
-#include "FireDomain.h"
-#include "include/Futils.h"
+#include "../FluxModel.h"
+#include "../FireDomain.h"
+#include "../include/Futils.h"
 using namespace std;
 
 namespace libforefire {
